feat: add Deno binary compilation
- Refactor index.ts into cli.ts (runtime-agnostic) + slim Node entrypoint
- Add Deno entrypoint with SDK Request patching for compatibility
- Replace Buffer/readFileSync with Uint8Array/TextDecoder/JSON imports
- Add deno.json with compile tasks for Linux and macOS binaries
- Add GitHub Actions workflow for building Deno binaries on release
- Add demo plugin and Deno Request patching unit tests
