+++ .github/workflows/deno-binaries.yml
@@ -1,0 +1,101 @@
+name: Deno Binaries
+
+on:
+  push:
+    tags:
+      - 'v*'
+  workflow_dispatch:
+
+jobs:
+  build-linux:
+    name: Build (Linux)
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v5
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v1
+        with:
+          deno-version: v2.x
+
+      - name: Slim package.json (prod deps only)
+        run: |
+          mv package.json package.full.json
+          deno eval "const p = JSON.parse(await Deno.readTextFile('package.full.json')); const out = { name: p.name, version: p.version, private: true, type: p.type, dependencies: p.dependencies }; await Deno.writeTextFile('package.json', JSON.stringify(out, null, 2) + '\n');"
+
+      - name: Compile
+        run: |
+          deno task compile:linux-x64
+          deno task compile:linux-arm64
+
+      - name: Upload artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: deno-binaries-linux
+          path: dist-deno/*
+
+  build-macos:
+    name: Build (macOS)
+    runs-on: macos-latest
+    permissions:
+      contents: read
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v5
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v1
+        with:
+          deno-version: v2.x
+
+      - name: Slim package.json (prod deps only)
+        run: |
+          mv package.json package.full.json
+          deno eval "const p = JSON.parse(await Deno.readTextFile('package.full.json')); const out = { name: p.name, version: p.version, private: true, type: p.type, dependencies: p.dependencies }; await Deno.writeTextFile('package.json', JSON.stringify(out, null, 2) + '\n');"
+
+      - name: Compile
+        run: |
+          deno task compile:mac-x64
+          deno task compile:mac-arm64
+
+      - name: Upload artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: deno-binaries-macos
+          path: dist-deno/*
+
+  upload-release-assets:
+    name: Upload Release Assets
+    runs-on: ubuntu-latest
+    needs: [build-linux, build-macos]
+    if: startsWith(github.ref, 'refs/tags/')
+    permissions:
+      contents: write
+
+    steps:
+      - name: Download Linux artifacts
+        uses: actions/download-artifact@v4
+        with:
+          name: deno-binaries-linux
+          path: dist-deno
+
+      - name: Download macOS artifacts
+        uses: actions/download-artifact@v4
+        with:
+          name: deno-binaries-macos
+          path: dist-deno
+
+      - name: Generate checksums
+        run: |
+          cd dist-deno
+          sha256sum * > SHA256SUMS.txt
+
+      - name: Upload to GitHub Release
+        uses: softprops/action-gh-release@v2
+        with:
+          files: dist-deno/*
