+++ .gitignore
@@ -85,6 +85,9 @@ out
 .nuxt
 dist
 
+# Deno compiled binaries
+dist-deno/
+
 # Gatsby files
 .cache/
 # Comment in the public line in if your project uses Gatsby and not Next.js
