+++ demo-plugin/c8ctl-plugin.js
@@ -1,0 +1,16 @@
+// c8ctl-plugin.ts
+export const metadata = {
+  name: 'my-plugin',
+  description: 'My custom c8ctl plugin',
+  commands: {
+    analyze: {
+      description: 'Analyze BPMN processes'
+    }
+  }
+};
+
+export const commands = {
+  analyze: async (args) => {
+    console.log('Analyzing...', args);
+  }
+};
