+++ deno.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "lib": ["es2023", "dom"]
+  },
+  "tasks": {
+    "check": "deno check --no-lock src/deno.ts",
+
+    "compile:mac-x64": "mkdir -p dist-deno && deno compile --no-lock --no-check --allow-net --allow-env --allow-sys --allow-read --allow-write --allow-run --target x86_64-apple-darwin --output dist-deno/c8ctl-macos-x64 src/deno.ts",
+    "compile:mac-arm64": "mkdir -p dist-deno && deno compile --no-lock --no-check --allow-net --allow-env --allow-sys --allow-read --allow-write --allow-run --target aarch64-apple-darwin --output dist-deno/c8ctl-macos-arm64 src/deno.ts",
+    "compile:linux-x64": "mkdir -p dist-deno && deno compile --no-lock --no-check --allow-net --allow-env --allow-sys --allow-read --allow-write --allow-run --target x86_64-unknown-linux-gnu --output dist-deno/c8ctl-linux-x64 src/deno.ts",
+    "compile:linux-arm64": "mkdir -p dist-deno && deno compile --no-lock --no-check --allow-net --allow-env --allow-sys --allow-read --allow-write --allow-run --target aarch64-unknown-linux-gnu --output dist-deno/c8ctl-linux-arm64 src/deno.ts",
+
+    "compile:all": "deno task compile:mac-x64 && deno task compile:mac-arm64 && deno task compile:linux-x64 && deno task compile:linux-arm64"
+  }
+}
